+++ lib/ar_lazy_preload/contexts/auto_preload_context.rb
@@ -4,13 +4,13 @@ module ArLazyPreload
   module Contexts
     # This class is responsible for automatic association preloading
     class AutoPreloadContext < BaseContext
-      protected
-
-      def association_needs_preload?(_association_name)
+      def auto_preload?
         true
       end
 
-      def auto_preload
+      protected
+
+      def association_needs_preload?(_association_name)
         true
       end
     end